+++ config-templates/module_perun.php
@@ -272,4 +272,15 @@ $config = [
             ],
         ],
     ],
+
+    'warning_test_sp_config' => [
+        'header' => [
+            'en' => '<h3>Warning - service in test environment</h3>',
+            'cs' => '<h3>Varování - testovací služba</h3>'
+        ],
+        'text' => [
+            'en' => '<p>Service is in the test environment.<br class="spacer"/>Hit the continue button.</p>',
+            'cs' => '<p>Služba je v testovacím režimu.<br class="spacer"/>Pokračujte zmáčknutím tlačítka.</p>',
+        ],
+    ],
 ];